refactor(client)!: refactor `Files` into `FilesUpload`
