+++ sn_client/src/lib.rs
@@ -26,9 +26,9 @@ pub use self::{
     event::{ClientEvent, ClientEventsReceiver},
     faucet::{get_tokens_from_faucet, load_faucet_wallet_from_genesis_wallet},
     files::{
-        api::FilesApi,
         download::{FilesDownload, FilesDownloadEvent},
-        FileUploadEvent, Files, BATCH_SIZE, MAX_UPLOAD_RETRIES,
+        upload::{FileUploadEvent, FilesUpload},
+        FilesApi, BATCH_SIZE, MAX_UPLOAD_RETRIES,
     },
     register::ClientRegister,
     wallet::{send, WalletClient},
