+++ sn_node/tests/nodes_rewards.rs
@@ -15,7 +15,7 @@ use crate::common::{
 use assert_fs::TempDir;
 use bls::{PublicKey, SecretKey, PK_SIZE};
 use eyre::{eyre, Result};
-use sn_client::{Client, ClientEvent, Files, WalletClient};
+use sn_client::{Client, ClientEvent, FilesUpload, WalletClient};
 use sn_logging::LogBuilder;
 use sn_node::{NodeEvent, ROYALTY_TRANSFER_NOTIF_TOPIC};
 use sn_protocol::safenode_proto::{
@@ -52,9 +52,9 @@ async fn nodes_rewards_for_storing_chunks() -> Result<()> {
     let prev_rewards_balance = current_rewards_balance()?;
     println!("With {prev_rewards_balance:?} current balance, paying for {} random addresses... {chunks:?}", chunks.len());
 
-    let mut files = Files::new(files_api.clone()).set_show_holders(true);
-    files.upload_chunks(chunks).await?;
-    let storage_cost = files.get_upload_storage_cost();
+    let mut files_upload = FilesUpload::new(files_api.clone()).set_show_holders(true);
+    files_upload.upload_chunks(chunks).await?;
+    let storage_cost = files_upload.get_upload_storage_cost();
 
     println!("Paid {storage_cost:?} total rewards for the chunks");
 
@@ -114,10 +114,10 @@ async fn nodes_rewards_for_chunks_notifs_over_gossipsub() -> Result<()> {
 
     tracing::info!("Paying for {num_of_chunks} random addresses...");
     println!("Paying for {num_of_chunks} random addresses...");
-    let mut files = Files::new(files_api.clone()).set_show_holders(true);
-    files.upload_chunks(chunks).await?;
-    let storage_cost = files.get_upload_storage_cost();
-    let royalties_fees = files.get_upload_royalty_fees();
+    let mut files_upload = FilesUpload::new(files_api.clone()).set_show_holders(true);
+    files_upload.upload_chunks(chunks).await?;
+    let storage_cost = files_upload.get_upload_storage_cost();
+    let royalties_fees = files_upload.get_upload_royalty_fees();
 
     println!("Random chunks stored, paid {storage_cost}/{royalties_fees}");
 
@@ -209,10 +209,10 @@ async fn nodes_rewards_transfer_notifs_filter() -> Result<()> {
 
     let num_of_chunks = chunks.len();
     println!("Paying for {num_of_chunks} chunks");
-    let mut files = Files::new(files_api.clone()).set_show_holders(true);
-    files.upload_chunks(chunks).await?;
-    let storage_cost = files.get_upload_storage_cost();
-    let royalties_fees = files.get_upload_royalty_fees();
+    let mut files_upload = FilesUpload::new(files_api.clone()).set_show_holders(true);
+    files_upload.upload_chunks(chunks).await?;
+    let storage_cost = files_upload.get_upload_storage_cost();
+    let royalties_fees = files_upload.get_upload_royalty_fees();
 
     println!("Random chunks stored, paid {storage_cost}/{royalties_fees}");
